[Updated the Brains of all the examples but did not retrain the environments]
