+++ unity-environment/Assets/ML-Agents/Examples/Soccer/Scripts/AgentSoccer.cs
@@ -94,17 +94,17 @@ public class AgentSoccer : Agent
             kickPower = 0f;
             switch (action)
             {
-                case 0:
+                case 1:
                     dirToGo = transform.forward * 1f;
                     kickPower = 1f;
                     break;
-                case 1:
+                case 2:
                     dirToGo = transform.forward * -1f;
                     break;
-                case 3:
+                case 4:
                     dirToGo = transform.right * -1f;
                     break;
-                case 2:
+                case 3:
                     dirToGo = transform.right * 1f;
                     break;
             }
@@ -114,23 +114,23 @@ public class AgentSoccer : Agent
             kickPower = 0f;
             switch (action)
             {
-                case 0:
+                case 1:
                     dirToGo = transform.forward * 1f;
                     kickPower = 1f;
                     break;
-                case 1:
+                case 2:
                     dirToGo = transform.forward * -1f;
                     break;
-                case 2:
+                case 3:
                     rotateDir = transform.up * 1f;
                     break;
-                case 3:
+                case 4:
                     rotateDir = transform.up * -1f;
                     break;
-                case 4:
+                case 5:
                     dirToGo = transform.right * -0.75f;
                     break;
-                case 5:
+                case 6:
                     dirToGo = transform.right * 0.75f;
                     break;
             }
