--- unity-environment/Assets/WallJumpAgentModified.cs.meta
@@ -1,13 +1,0 @@
-fileFormatVersion: 2
-guid: 6fbc57e4fadbf4086923b4fef6067462
-timeCreated: 1517445814
-licenseType: Free
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
