--- unity-environment/Assets/WallJumpModified.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 82cb195e372814689961fbe6c689f156
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
